handling DB lock issue for inspection template save

+++ src/main/java/com/facilio/qa/command/DeleteInspectionTriggersCommand.java
@@ -15,6 +15,7 @@ import com.facilio.bmsconsoleV3.context.inspection.InspectionTriggerContext;
 import com.facilio.bmsconsoleV3.util.InspectionAPI;
 import com.facilio.chain.FacilioContext;
 import com.facilio.constants.FacilioConstants;
+import com.facilio.db.builder.GenericDeleteRecordBuilder;
 import com.facilio.db.criteria.CriteriaAPI;
 import com.facilio.db.criteria.operators.DateOperators;
 import com.facilio.db.criteria.operators.EnumOperators;
@@ -22,6 +23,7 @@ import com.facilio.db.criteria.operators.NumberOperators;
 import com.facilio.fw.BeanFactory;
 import com.facilio.modules.DeleteRecordBuilder;
 import com.facilio.modules.FieldFactory;
+import com.facilio.modules.ModuleFactory;
 import com.facilio.modules.fields.FacilioField;
 import com.facilio.time.DateTimeUtil;
 import com.facilio.v3.context.Constants;
@@ -45,10 +47,30 @@ public class DeleteInspectionTriggersCommand extends FacilioCommand {
 		
 		Map<String, FacilioField> fieldMap = FieldFactory.getAsMap(fields);
 		
-		DeleteRecordBuilder<InspectionTriggerContext> deleteBuilder = new DeleteRecordBuilder<InspectionTriggerContext>()
-				.module(modBean.getModule(FacilioConstants.Inspection.INSPECTION_TRIGGER))
-				.andCondition(CriteriaAPI.getCondition(fieldMap.get("parent"), inspectionIds, NumberOperators.EQUALS));
-		deleteBuilder.delete();
+		List<InspectionTriggerContext> triggers = InspectionAPI.getInspectionTrigger(CriteriaAPI.getCondition(fieldMap.get("parent"), inspectionIds, NumberOperators.EQUALS), false);
+		
+		if(triggers != null && triggers.size() > 0) {
+			
+			List<Long> triggerIds = triggers.stream().map(InspectionTriggerContext::getId).collect(Collectors.toList());
+			
+			DeleteRecordBuilder<InspectionTriggerContext> deleteBuilder = new DeleteRecordBuilder<InspectionTriggerContext>()
+					.module(modBean.getModule(FacilioConstants.Inspection.INSPECTION_TRIGGER))
+					.andCondition(CriteriaAPI.getIdCondition(triggerIds, modBean.getModule(FacilioConstants.Inspection.INSPECTION_TRIGGER)));
+			deleteBuilder.delete();
+			
+			List<Long> scheduleIds = triggers.stream().filter(t -> (t.getScheduleId() != null && t.getScheduleId() > 0)).map(InspectionTriggerContext::getScheduleId).collect(Collectors.toList());
+			
+			if(scheduleIds != null && scheduleIds.size() > 0) {
+				
+				GenericDeleteRecordBuilder delete = new GenericDeleteRecordBuilder()
+						.table(ModuleFactory.getBaseSchedulerModule().getTableName())
+						.andCondition(CriteriaAPI.getIdCondition(scheduleIds, ModuleFactory.getBaseSchedulerModule()))
+						;
+				
+				delete.delete();
+			}
+			
+		}
 		
 		InspectionAPI.deleteScheduledPreOpenInspections(inspectionIds);
